# 🚀 Complete Migration from .NET Framework 4.8 to .NET 8