--- a/TaxCalculator.AspNetCore.Api.Tests/TaxCalculator.AspNetCore.Api.Tests.csproj
+++ b/TaxCalculator.AspNetCore.Api.Tests/TaxCalculator.AspNetCore.Api.Tests.csproj
@@ -1,25 +1,24 @@
-<Project Sdk="Microsoft.NET.Sdk">
+<Project Sdk="Microsoft.NET.Sdk.Web">
 
   <PropertyGroup>
     <TargetFramework>net8.0</TargetFramework>
+    <Nullable>enable</Nullable>
     <ImplicitUsings>enable</ImplicitUsings>
-    <Nullable>disable</Nullable>
-    <GenerateAssemblyInfo>false</GenerateAssemblyInfo>
     <IsPackable>false</IsPackable>
     <IsTestProject>true</IsTestProject>
   </PropertyGroup>
 
   <ItemGroup>
+    <PackageReference Include="Microsoft.AspNetCore.Mvc.Testing" Version="8.0.0" />
     <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.8.0" />
     <PackageReference Include="NUnit" Version="3.14.0" />
     <PackageReference Include="NUnit3TestAdapter" Version="4.5.0" />
-    <PackageReference Include="Moq" Version="4.20.70" />
+
   </ItemGroup>
 
   <ItemGroup>
+    <ProjectReference Include="..\TaxCalculator.AspNetCore.Api\TaxCalculator.AspNetCore.Api.csproj" />
     <ProjectReference Include="..\TaxCalculator.Core\TaxCalculator.Core.csproj" />
-    <ProjectReference Include="..\TaxCalculator.Data\TaxCalculator.Data.csproj" />
-    <ProjectReference Include="..\TaxCalculator.Services\TaxCalculator.Services.csproj" />
   </ItemGroup>
 
 </Project>