--- a/TaxCalculator.AspNetCore.Api.Tests/TaxCalculator.AspNetCore.Api.Tests.csproj
+++ b/TaxCalculator.AspNetCore.Api.Tests/TaxCalculator.AspNetCore.Api.Tests.csproj
@@ -11,6 +11,7 @@
   <ItemGroup>
     <PackageReference Include="Microsoft.AspNetCore.Mvc.Testing" Version="8.0.0" />
     <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.8.0" />
+    <PackageReference Include="Moq" Version="4.20.70" />
     <PackageReference Include="NUnit" Version="3.14.0" />
     <PackageReference Include="NUnit3TestAdapter" Version="4.5.0" />
 
@@ -19,6 +20,8 @@
   <ItemGroup>
     <ProjectReference Include="..\TaxCalculator.AspNetCore.Api\TaxCalculator.AspNetCore.Api.csproj" />
     <ProjectReference Include="..\TaxCalculator.Core\TaxCalculator.Core.csproj" />
+    <ProjectReference Include="..\TaxCalculator.Data\TaxCalculator.Data.csproj" />
+    <ProjectReference Include="..\TaxCalculator.Services\TaxCalculator.Services.csproj" />
   </ItemGroup>
 
 </Project>