--- a/TaxCalculator.AspNetCore.Api/TaxCalculator.AspNetCore.Api.http
+++ b/TaxCalculator.AspNetCore.Api/TaxCalculator.AspNetCore.Api.http
@@ -1,0 +1,6 @@
+@TaxCalculator.AspNetCore.Api_HostAddress = http://localhost:5256
+
+GET {{TaxCalculator.AspNetCore.Api_HostAddress}}/weatherforecast/
+Accept: application/json
+
+###