--- a/TaxCalculator.AspNetCore.Api/appsettings.Development.json
+++ b/TaxCalculator.AspNetCore.Api/appsettings.Development.json
@@ -1,0 +1,8 @@
+{
+  "Logging": {
+    "LogLevel": {
+      "Default": "Information",
+      "Microsoft.AspNetCore": "Warning"
+    }
+  }
+}