--- a/TaxCalculator.Data/Repositories/SqlConnectionFactory.cs
+++ b/TaxCalculator.Data/Repositories/SqlConnectionFactory.cs
@@ -1,5 +1,5 @@
 using System.Data;
-using System.Data.SqlClient;
+using Microsoft.Data.SqlClient;
 using TaxCalculator.Data.Interfaces;
 
 namespace TaxCalculator.Data.Repositories