--- a/TaxCalculator.Data/Repositories/UserIncomeRepository.cs
+++ b/TaxCalculator.Data/Repositories/UserIncomeRepository.cs
@@ -1,6 +1,6 @@
 using System;
 using System.Collections.Generic;
-using System.Data.SqlClient;
+using Microsoft.Data.SqlClient;
 using System.Threading.Tasks;
 using TaxCalculator.Core.Models;
 using TaxCalculator.Data.Interfaces;