--- a/TaxCalculator.Data/Repositories/UserRepository.cs
+++ b/TaxCalculator.Data/Repositories/UserRepository.cs
@@ -1,6 +1,6 @@
 using System;
 using System.Collections.Generic;
-using System.Data.SqlClient;
+using Microsoft.Data.SqlClient;
 using System.Threading.Tasks;
 using TaxCalculator.Core.Models;
 using TaxCalculator.Data.Interfaces;