… .github/workflows/dotnet8-ci.yml
… CICD-Migration-Plan.md
… DotNet8-Migration-Analysis.md
… MIGRATION-COMPLETE.md
… PR_DESCRIPTION.md
… README.md
… TaxCalculator.AspNetCore.Api/appsettings.json
… TaxCalculator.Infrastructure/TaxCalculator.Infrastructure.csproj
… TaxCalculator.Services/Services/TaxCalculationService.cs
… TaxCalculator.Services/TaxCalculator.Services.csproj
… progress.md
… specs/migration-overview.md
… specs/phase-1.md
… specs/phase-2.md
… specs/phase-3.md
… specs/phase-4.md
… specs/risk-assessment.md